Moved calendar invite components into the calendar-invite.tsx file
