+++ frontend/cotangles/src/components/my-calendars/calendar-list.tsx
@@ -1,9 +1,8 @@
 import * as React from "react";
 import { Button } from "../ui/button";
-import { Check, X } from "lucide-react";
 import { useNavigate } from "react-router-dom";
 import { cn } from "@/lib/utils";
-import { CalendarCreateForm } from "./calendar-create-form";
+import { CardCalendarProps } from "./calendar-invite";
 
 
 // Components specifically related to the my calendar page
@@ -23,43 +22,6 @@ const CardBodyCalendar = React.forwardRef<HTMLDivElement, React.HTMLAttributes<H
   );
   CardBodyCalendar.displayName = "CardBodyCalendar";
 
-// Todo, link this to back end , below component will have the invites prop
-// Todo: add a state that removes pending invites when either accept or deny button is clicked
-const CardCalendarInfo = () => {
-    return (
-      <div className="flex flex-col content-center items-center justify-between rounded-[2.5em] bg-transparent text-black w-[30%] h-[70%]">
-        <CalendarCreateForm/>
-          <div className="flex flex-col items-center justify-start rounded-[2.5em] shadow-light border-2 border-border bg-secondary text-black w-full h-[70%] overflow-y-auto">
-            <h1 className="break-words w-full text-md py-2">Pending</h1>
-            <CalendarInvite calendarId={1} calendarName="This is20 characters"/>
-            <CalendarInvite calendarId={2} calendarName="omg"/>
-            <CalendarInvite calendarId={3} calendarName="Gigachad meetups"/>
-            <CalendarInvite calendarId={3} calendarName="Gigachad meetups"/>
-            <CalendarInvite calendarId={3} calendarName="Gigachad meetups"/>
-            <CalendarInvite calendarId={3} calendarName="Gigachad meetups"/>
-            <CalendarInvite calendarId={3} calendarName="Gigachad meetups"/>
-          </div>
-      </div>
-    );
-  };
-
-  
-type CardCalendarProps = { calendarName: string, calendarId: number }
-
-const CalendarInvite = ({ calendarName, calendarId }: CardCalendarProps) => {
-    // Adds the calendarId to + name as a card to the my-calendars page when you
-    // accept
-    // Removes the invite when you deny
-    console.log(calendarId)
-    return(
-        <div className="flex justify-center content-center items-center w-[90%] p-0 pb-1">
-            <span className="flex flex-grow h-[80%] min-w-[60%] text-[0.7em] bg-primary text-center justify-center items-center border-2 rounded-lg px-1"><b>{calendarName}</b></span>
-            <Button className="scale-75 w-[20%] rounded-xl bg-affirmative h-full"><Check/></Button>
-            <Button className="scale-75 w-[20%] rounded-xl bg-destructive h-full"><X/></Button>
-        </div>
-    )
-}
-
 const CardCalendar = ({ calendarName, calendarId }: CardCalendarProps) => {
     const navigate = useNavigate()
     return (
@@ -77,6 +39,5 @@ const CardCalendar = ({ calendarName, calendarId }: CardCalendarProps) => {
 
 export {
     CardCalendar,
-    CardBodyCalendar,
-    CardCalendarInfo
+    CardBodyCalendar
 }
